OS Notification on update dowloaded and minor refactor
