--- preview/helper/notify.js
@@ -1,9 +1,0 @@
-function Notify(options) {
-  const{ title, body }  = options;
-  return new Notification(title, {
-    body,
-  });
-}
-
-module.exports = { Notify };
-
